Check @Plugins for prereqs at "make install" time
We cannot check @Plugins earlier, such as during "perl Makefile.PL"
time, because of permissions.  "make install", however, does have
permissions to read the RT_Config file.  The set of prereqs cannot be
hardcoded into the call in the Makefile because they are not known yet;
only calls to rt_requires_plugin() after the RTx() call will set them
up.  Thus the only source of the RT plugin prereqs is in the META.yml
file; load it and to check that all of the explicit dependencies are
enabled.  This mirrors existing code in inc::Module::Install::Metadata.

The requires_rt line was moved earlier because it also sets a minimum
version of perl, which is necessary for ->include_deps to determine what
needs to be bundled along with YAML::Tiny in the distribution.