Switch Module::Install::RTx::Runtime to no longer inherit from M::I
Inheriting from Module::Install means that, in the presence of a
inc/.author file, a BEGIN block is run that clears out inc/.  This on
purpose, such that the updated Module::Install files can be copied in.
However, in the case of the runtime parts of Module::Install::RTx, this
meant that running "make initdb" (and now "make install") would clear
our inc/ if run as an author -- and replace them with the much smaller
set of dependencies that the runtime module needed.

Make Module::Install::RTx::Runtime truely runtime-only, by importing it
explicitly into inc/, and making it no longer inherit from
Module::Install.