Make rt_too_new warn instead to not block installation
E.g.

    rt_too_new('5.2.0');

This usually doesn't mean the extension won't work with RT 5.2(since
it's not released yet, we actually don't know if it will work or not for
now). It's quite possible that current extension works fine with RT 5.2
without any changes.

Thus here we choose to warn that it's not tested yet.