Add red color to warnings to make them stand out
Module::Install prints a bunch of information to the terminal
when you run Makefile.PL and that can make it easy for
installers to miss warning messages that they should read. Add
text color to make messages stand out.

The \n is moved to the end of the warn lines to suppress printing
the "at linenumber" messages in output.